app/main/services/speech_recognition/requirements.txt
deepspeech-gpu==0.9.3
Flask==1.1.4
flask-restx==0.4.0